Common: Move PluginType enum from types.py to plugin/plugin_type.py
