+++ monkey/common/plugins/plugin_type.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class PluginType(Enum):
+    CREDENTIAL_COLLECTOR = "CredentialCollector"
+    EXPLOITER = "Exploiter"
+    FINGERPRINTER = "Fingerprinter"
+    PAYLOAD = "Payload"
