+++ monkey/infection_monkey/i_puppet/i_puppet.py
@@ -5,7 +5,8 @@ from dataclasses import dataclass
 from typing import Dict, Mapping, Optional, Sequence
 
 from common.credentials import Credentials
-from common.types import PingScanData, PluginType
+from common.plugins.plugin_type import PluginType
+from common.types import PingScanData
 from infection_monkey.model import VictimHost
