+++ monkey/infection_monkey/puppet/puppet.py
@@ -5,7 +5,8 @@ from typing import Dict, Sequence
 from common.common_consts.timeouts import CONNECTION_TIMEOUT
 from common.credentials import Credentials
 from common.event_queue import IAgentEventQueue
-from common.types import PingScanData, PluginType
+from common.plugins.plugin_type import PluginType
+from common.types import PingScanData
 from infection_monkey import network_scanning
 from infection_monkey.i_puppet import ExploiterResultData, FingerprintData, IPuppet, PortScanData
 from infection_monkey.master.plugin_registry import PluginRegistry
