+++ monkey/tests/unit_tests/infection_monkey/puppet/test_puppet.py
@@ -2,7 +2,8 @@ import threading
 from unittest.mock import MagicMock
 
 from common.event_queue import IAgentEventQueue
-from common.types import PingScanData, PluginType
+from common.plugins.plugin_type import PluginType
+from common.types import PingScanData
 from infection_monkey.master.plugin_registry import PluginRegistry
 from infection_monkey.puppet.puppet import EMPTY_FINGERPRINT, Puppet
